Improved regex flag support.
Each Ovid processor now takes a re module ‘flags’ argument to the
initializer, and this is passed to re.compile for that processor.
There is likewise a new ‘flags’ property on DelimitedShorthand to
serve in the compilation of class-wide regexes. The latter change
replaces the practice of passing a ‘flags’ argument to
collective_sub(). Any such argument will now be ignored in silence.

Taken together, these changes allow for easy multi-line shorthand.